Move resource pkg at server root
